Create README.md
The purpose of the app is to provide bitesize chunks of information
about various philosophers of the user's choosing. The app has two buttons - one of which shuffles various quotes attributed to the philosopher in
question, the other of which shuffles various pieces of biographical information regarding the philosopher. The app is designed to mimic those decks of 
revision cards often used by secondary school students, who want a portable means of preparing for exams while on the move (though this
is geared more towards adult autodidacts). At the moment, I've only added a small number of philosophers for demonstration purposes but the .txt
files from which the app derives its information could easily be built out. User passwords and names are stored in a json file.